requirements.txt
flask
iptools
requests
waitress
gitpython
pyopenssl